chore: rename legacy badges to follow the new naming convention
