+++ ui/components/icons/providers-badge/aws-provider-badge.tsx
