+++ ui/components/icons/providers-badge/azure-provider-badge.tsx
