+++ ui/components/icons/providers-badge/gcp-provider-badge.tsx
