+++ ui/components/icons/providers-badge/index.ts
@@ -1,4 +1,4 @@
-export * from "./AWSProviderBadge";
-export * from "./AzureProviderBadge";
-export * from "./GCPProviderBadge";
-export * from "./KS8ProviderBadge";
+export * from "./aws-provider-badge";
+export * from "./azure-provider-badge";
+export * from "./gcp-provider-badge";
+export * from "./ks8-provider-badge";
